bruuh

+++ Content.Shared/_Scp/Other/ScpOnSoundVisibility/ScpOnSoundVisibilityComponent.cs
@@ -5,17 +5,17 @@ public sealed partial class ScpOnSoundVisibilityComponent : Component
 {
     public const float InitialVisibilityAcc = 0.001f;
     public const float DefaultHideTime = 2.5f;
-    public const int DefaultMinValue = 40;
-    public const int DefaultMaxValue = 400;
+    public const float DefaultMinValue = 40f;
+    public const float DefaultMaxValue = 400f;
 
     [DataField]
     public float HideTime = DefaultHideTime;
 
     [DataField]
-    public int MinValue = DefaultMinValue;
+    public float MinValue = DefaultMinValue;
 
     [DataField]
-    public int MaxValue = DefaultMaxValue;
+    public float MaxValue = DefaultMaxValue;
 
     [DataField]
     public float SharePhraseRadius = 16f;
